+++ Wati_PushSwap.h
@@ -6,7 +6,7 @@
 /*   By: tschlege <tschlege@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/16 19:57:05 by tschlege          #+#    #+#             */
-/*   Updated: 2022/07/14 18:41:28 by tschlege         ###   ########lyon.fr   */
+/*   Updated: 2022/07/14 20:35:29 by tschlege         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,34 +25,33 @@ typedef struct s_stack {
 	int	counter;
 }	t_stack;
 
-void	ft_putstr_fd(char	*s, int fd);
-void	wati_exit(char *str, int choice);
-void	wati_parser(char **argv, t_stack *stack);
-int		ft_isdigit(int c);
-int		ft_atoi(const char *str);
-void	ft_putnbr_to_stack_a(char **argv, t_stack *stack);
-void	hardcode_sort(t_stack *stack);
-int		a_is_sorted(t_stack *stack);
-void	radix(t_stack *stack);
-int		one_shot_action(t_stack *stack);
-
+void			ft_putstr_fd(char	*s, int fd);
+void			wati_exit(char *str, int choice);
+void			wati_parser(char **argv, t_stack *stack);
+int				ft_isdigit(int c);
+long long int	ft_atoi(const char *str);
+void			ft_putnbr_to_stack_a(char **argv, t_stack *stack);
+void			hardcode_sort(t_stack *stack);
+int				a_is_sorted(t_stack *stack);
+void			radix(t_stack *stack);
+int				one_shot_action(t_stack *stack);
 /*
 ** Wati Stacks Actions 
 */
-int		wati_sa(t_stack *stack);
-int		wati_sb(t_stack *stack);
-int		wati_ss(t_stack *stack);
-int		wati_ra(t_stack *stack);
-int		wati_rb(t_stack *stack);
-void	wati_rr(t_stack *stack);
-int		wati_rra(t_stack *stack);
-int		wati_rrb(t_stack *stack);
-void	wati_rrr(t_stack *stack);
-int		wati_pa(t_stack *stack);
-int		wati_pb(t_stack *stack);
+int				wati_sa(t_stack *stack);
+int				wati_sb(t_stack *stack);
+int				wati_ss(t_stack *stack);
+int				wati_ra(t_stack *stack);
+int				wati_rb(t_stack *stack);
+void			wati_rr(t_stack *stack);
+int				wati_rra(t_stack *stack);
+int				wati_rrb(t_stack *stack);
+void			wati_rrr(t_stack *stack);
+int				wati_pa(t_stack *stack);
+int				wati_pb(t_stack *stack);
 /*
 ** for tests 
 */
-void	print_stacks(t_stack *stack);
-void	debug(t_stack *stack);
+void			print_stacks(t_stack *stack);
+void			debug(t_stack *stack);
 #endif
